Add update_stock_data view
The /update_stock_data url loads the update_stock_data view that begins fetching
the stock details for the SELECTED_STOCK_TICKERS defined in
settings.py and adding the stock information to the postgresql database.

Currently, this page will only work to populate the database initially.
The view does not yet support updating existing stock data.
This will have to be worked on eventually.

Navigating to the /update_stock_data URL will take some time to load since the
page does not load until it is done updating all the stock data.
The script also has to sleep for a minute between each stock to avoid reaching
the 5 call/min API limit.