Add /stocks/<data>/<ticker> endpoints
Add endpoints for Daily Adjusted, SMA, VWAP, and RSI data.
This data for the available stocks can be queried.
The only query parameters used at the moment are start_time and end_time which
specify the range of data to retrieve.

To query this data, the endpoint is /stocks/<data>/<ticker> i.e. for getting
SMA data for TSLA, the endpoint is /stocks/SMA/TSLA.

The following endpoints were added:
/stocks/daily_adjusted/<ticker>
/stocks/SMA/<ticker>
/stocks/VWAP/<ticker>
/stocks/RSI/<ticker>

Since the data returned by each endpint is so similar, the stock_data_get
decorator simplifies the processing of each request.

See the documentation in the code for the request/response format for each
endpoint.